+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJF6K6SPY8EGHWC62JYWRSHN"
+        "01KJFCKVCMPKSRKMQXAA96WRTW"
     )
     r"""The unique identifier of the data source"""