+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJF6K6STN95KTJFVVWWCMS07"
+        "01KJFCKVCWN6F2E7DHT0QG45HG"
     )
     r"""The unique identifier of the data source"""